Refactor the server logic
Wrap the server logic in a class called HTTPServer. serve_forever sets up the server socket and accepts new connections on a loop, it then sends those connections over to handle_client which processes the request and sends a response over the socket.

Define a HTTPWorker class, HTTP workers are OS threads that wait for new connections to pop up on a queue and then act on them.